feat: implement correctly the sections selection
